CommandLine: Separate accept/parse phases
Rework internals so acceptXxx() only registers known options, then
parseOptions(args) performs the actual parsing.  The previous approach
exhibited unintended behavior, like:

    -a -b c

Given -a requires an argument and -b is a no-arg option (a flag), if
-b is accepted before -a, there won't be an error and -a will get an
argument of "c".  If -a is accepted before -b, -a will get an argument
of "-b" (not yet known), etc.