SmoothDownscale: Reverse scaling strategy
First scale to a size that is integral factor of the target dimensions,
and is greater than or equal to half the source size.  Then perform as
many half downscale operations necessary to get to the target size.

Subjectively this produces better results, with more perceived detail
preserved.  It could be just me imagining, also.

Have SmoothDownscale.resize() produce the target size, already.