XCursor: Ensure bitmap width >= nominal size
Vertical cropping occurs when the bitmap width for a greater nominal
size is close to a lesser nominal size.  Especially visible with the
xterm (I-beam) pointer.

Crop the bitmap height to the nominal size when not cropping to content
(not an option, currently).

IntPixels.cropTo -> resizeCanvas - allow expanding the source canvas,
if ever comes to play.