XCursor: Ensure bitmap height >= nominal size
I'm observing random pixels at the bottom of the displayed pointers,
for bitmaps with height < the nominal size - very noticeable with the
`vertical-text` pointer.  I'm still not sure about the requirements:
nominal size vs. actual bitmap size, and the problem seemingly goes away
when the display pointer size is increased above certain threshold
(dconf Editor):

    org/gnome/desktop/interface/cursor-size

F.e. the "broken pixels" appear with the default size 24, and a size of
32 but don't appear with size 40 or greater.

It could be just a quirk of the VirtualBox VM I'm testing with as I
notice other anomalies with it, even when I force the bitmap sizes to
uniform (square) nominal size:

-   Animations don't work with display size < 40
-   For display size < 40, the bitmaps are rendered as if ignoring the
    fact they (should) have the alpha premultiplied - darker stroke
    visible at the edges of white outlines (classic theme cursors)

Amends 3abbbf1114de19aaf902d6045e4417aa495fab6b