SVGTransformer: Implement transform pipeline
Make SVGTransformer apply the drop-shadow and svg11-compat transformations
consecutively, the latter on the result of the former.  This is possibly
slower but allows the transformations to be independent from each other, and
allows for new transformations to be easily added (though should be avoided
for the mentioned performance penalty).

To mitigate the performance slow down an efficient in-memory SAXReplayBuffer
is introduced.  It gets utilized in the existing SVGSizing (when loading/saving
to file), and ColorWheel.

Additionally:

-   SVGCursorMetadata: allow reading from any TrAX source
-   svg11-compat: fix @href rewriting rule not matching