SVGTransformer: Add possibility of applying "thin-stroke" transformation
Currently not exposed in the CLI or API (off by default).  To be fully usable
alignment anchors need to specify appropriate bias and be adjusted with offset
of (base-width - new-width) / 2.  Also it is possibly applicable to the hotspots
but not as important as for the alignment hints.

I'm unclear whether I'll keep this as XSLT or possibly making a DOM utility
like ColorScheme.