CursorNames: Load mapping dynamically
    --windows-cursors

implies --windows-cursors=win-names.json, if "win-names.json" is present in
the project directory.

    --linux-cursors

implies --linux-cursors=x11-names.json, if "x11-names.json" is present in
the project directory.

The mapping also acts as a filter to the source files processed.  The existing
additional [-f <cursor>]... filter options are applied to the name mapping, if
available, or act as <cursor>=<cursor> mapping (and filter).