BitmapsRenderer: Introduce --default-stroke-also and --no-shadow-also options
Allow f.e. only thin-stroke but variants with shadow and no shadow:

    --pointer-shadow --no-shadow-also --thin-stroke

or the other way around – variants with thin stroke and the default stroke,
and all variants have shadow:

    --thin-stroke --default-stroke-also --pointer-shadow

As previously, --all-variants is a shorthand for:

    --no-shadow-also --default-stroke-also

Note, there's no --default-color-also.  Define empty mapping in "colors.json":

    "Default": {}

and specify it explicitly:

    --color Black,White,Default