AnchorPoint: Allow specifying bias in any direction (angle) and distance
The "top", "bottom", "left", and "right" tokens imply relative -1, 1, -1, and 1
direction accordingly in vertical, and horizontal direction.  It is now possible:

    bias-left-t221

meaning dX=-1, dY=-2.21.  Use the first letter from "top", "bottom", "left", and
"right" followed by an integer converted to a float with a fixed decimal point
after the first digit, or with at least two decimal digits:

    t5 -> dY=-0.05
    b25 -> dY=0.25
    l125 -> dX=-1.25
    r2345 -> dX=2.345

Added a 'mode' property controlling when the bias should be applied.  To be
used further with an "expand-fill" stroke thinning mode.