BitmapsRenderer: --expand-fill[=<base-width>] argument
Add optional argument to --expand-fill to control the amount the fill is
expanded:

    --stroke-width=8 --expand-fill=12

The fill will be expanded as if the base stroke width was 12, instead of
the implied --base-stroke-width=16.  That is, it will limit the amount the
fill is expanded.  It is possible to expand the fill beyond to original
stroke, but that would produce likely less practical results:

    --stroke-width=8 --expand-fill=20

The actual base stroke width is still used/necessary to calculate the amount
the fill needs to be expanded, and the stroke to be shrunk (or expanded).

---

The used naming (class names):

    bias-*
    bias-*-reverse
    bias-*-always

    fixed-width-stroke
    expand-fill-stroke
    outer-stroke
    stroke-only
    shrink-only-stroke

could/should likely be improved, and possibly expanded.