BitmapsRenderer: Change --expand-fill[=<limit>] meaning
Change the meaning of the --expand-fill argument to express a limit to
expansion of the fill.  W/o the argument the behavior is the same.

It is now possible to:

  --stroke-width=12 --stroke-width=8 --expand-limit=4

Previously, to simulate the expand-fill limit with the second stroke-width:

  --stroke-width=12 --stroke-width=8 --expand-limit=12

would produce just "shrink stroke" result with the first stroke-width.

It is no longer possible to expand the fill beyond the base-stroke-width - it
doesn't have practical application.