thin-stroke: Compensate expanding fill stroke at low resolutions
At low resolutions the expanding stroke could be < 1 target pixel and its
anti-aliasing over the anti-aliasing of the edge of the fill may not fully
"cover" that pixel – would not result in a solid fill to the outer edge of
the stroke, and may look like misalignment.  Add one or more layers of fill
to partially compensate (cover "the crack").