SVGSizing: Keep the hotspot "inside" the pointer shape
If hotspot anchor has a positive bias, the coordinate is deemed to the outer
edge of the shape, so subtract 1 from the rounded coordinate to keep it inside
the shape.  This is modeled after existing behavior - drawing a line or a
rectangle with (x1,y1) and (x2,y2) the last filled point is just before (x2,y2).

If there's no bias specified, decrement only if rounded bigger.  This appears to
achieve better results with the default center hotspot (f.e. see the "cross" cursor).

Optionally (via system property configuration) use the source view-box center to
determine coordinate "on the outside".