CursorRenderer: Extract BitmapOutputBuilder
Simplify the implementation by eliminating conditional branching explicitly
and only for handling OutputType.BITMAP.  As a side effect the RendererBackend
interface has also been simplified - no writeXxx methods.  The latter are
not performance critical as BITMAP output is considered a debug mode.  The
performance can't be any better with the default (JSVG) backend in this
mode, also.