index.theme: Remove "hicolor" quotes
While the specification [1] says:

>   If no theme is specified, implementations are required to add the "hicolor"
>   theme to the inheritance tree.

the examples [2] given incl. in the real-world use no quotes:

    Inherits=wood,default

This is also according to the general format:

-   https://specifications.freedesktop.org/desktop-entry-spec/latest/basic-format.html#entries
-   https://specifications.freedesktop.org/desktop-entry-spec/latest/value-types.html

Note `Inherits=hicolor` should generally not need to be specified explicitly,
but I'm seeing it often enough to presume there may be an obscure compatibility
reason to do so.

[1] https://specifications.freedesktop.org/icon-theme-spec/latest/#file_formats
[2] https://specifications.freedesktop.org/icon-theme-spec/latest/#example