MouseGen: Add --config <path> option
Allow to specify a --config path to resolve implicit/default configuration
files against.  It would usually be relative to the <project-dir> like:

    <project-dir>/
      configs/
        animations.json
        colors.json
        win-names.json
        x11-names.json

So:

    --config configs
    --windows-cursors

instead of:

    --animations configs/animations.json
    --color-map configs/colors.json
    --windows-cursors=configs/win-names.json

"render.json" is still implicitly looked up in the <project-dir> but could
also be specified like:

    --config configs/render1.json

The base config directory becomes "configs", but note paths specified
in the "render.json" and --source dirs are still resolved against the
effective <project-dir>.

The behavior of the following is currently undefined:

    mousegen <project-dir>/renderA.json --config <config-dir>/renderB.json

"render.json" should be specified only once.