WIP: ScalableBuilder
Need to reuse much of the CursorRenderer preprocessing.  The interface
is very similar to CursorBuilder but it doesn't use RendererBackend to
produce bitmaps and then the builder needs the DOM Document vs. a
BufferedImage.  Need to think how/if I could unify this.

Outputs KDE Plasma format:

-   https://github.com/KDE/breeze/tree/v6.4.91/cursors/Breeze/Breeze/cursors_scalable/default
-   https://blogs.kde.org/2025/01/12/kcursorgen-and-svg-cursors/

hyprcursor appears more complicated (for starters):

-   https://standards.hyprland.org/hyprcursor/